+++ Classes/Db.php
@@ -739,7 +739,7 @@ class Tx_Oelib_Db
     /**
      * Returns $GLOBALS['TYPO3_DB'].
      *
-     * @deprecated will be removed in oelib 3.0
+     * @deprecated will be removed in oelib 4.0
      *
      * @return DatabaseConnection
      */