+++ Classes/Domain/Repository/Traits/ReadOnlyTrait.php
@@ -5,7 +5,7 @@ namespace OliverKlee\Oelib\Domain\Repository\Traits;
 /**
  * This trait marks repositories as read-only.
  *
- * @deprecated Will be removed in oelib 3.0.0. Use ReadOnly (without the "Trait" suffix) instead.
+ * @deprecated Will be removed in oelib 3.0. Use ReadOnly (without the "Trait" suffix) instead.
  *
  * @author Oliver Klee <typo3-coding@oliverklee.de
  */